requirements.txt
flask
pandas
openpyxl
openai
python-dotenv
Pillow
python-docx
